requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-genai
google-auth
sse-starlette
httpx
requests
websockets
certifi
pydantic
